feat: create lighthouse.md
I've been thinking a lot about whether Carmel should just stick to debates, or if it would make sense to merge Lighthouse into it. Or perhaps keeping Lighthouse as a standalone subproject, as it is now, is the best way to go.

What I really feel though, is that both of these projects could benefit from a shared core framework and then become specialized pipelines on top of it.

Even the name is giving me pause.

But enough about what I think. I want to hear from you. What are your thoughts on all of this?